relocate SlotUtil
