+++ src/main/java/gregtech/api/gui/widgets/PhantomSlotUtil.java
@@ -1,10 +1,12 @@
-package gregtech.api.util;
+package gregtech.api.gui.widgets;
 
 import net.minecraft.inventory.ClickType;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
 
-public class SlotUtil {
+public final class PhantomSlotUtil {
+
+    private PhantomSlotUtil() {/**/}
 
     public static ItemStack slotClickPhantom(Slot slot, int mouseButton, ClickType clickTypeIn, ItemStack stackHeld) {
         ItemStack stack = ItemStack.EMPTY;
