+++ src/main/java/gregtech/api/gui/widgets/PhantomSlotWidget.java
@@ -2,7 +2,6 @@ package gregtech.api.gui.widgets;
 
 import com.google.common.collect.Lists;
 import gregtech.api.gui.ingredient.IGhostIngredientTarget;
-import gregtech.api.util.SlotUtil;
 import gregtech.client.utils.TooltipHelper;
 import mezz.jei.api.gui.IGhostIngredientHandler.Target;
 import net.minecraft.entity.player.EntityPlayer;
@@ -67,7 +66,7 @@ public class PhantomSlotWidget extends SlotWidget implements IGhostIngredientTar
     @Override
     public ItemStack slotClick(int dragType, ClickType clickTypeIn, EntityPlayer player) {
         ItemStack stackHeld = player.inventory.getItemStack();
-        return SlotUtil.slotClickPhantom(slotReference, dragType, clickTypeIn, stackHeld);
+        return PhantomSlotUtil.slotClickPhantom(slotReference, dragType, clickTypeIn, stackHeld);
     }
 
     @Override
@@ -94,7 +93,7 @@ public class PhantomSlotWidget extends SlotWidget implements IGhostIngredientTar
                     int mouseButton = Mouse.getEventButton();
                     boolean shiftDown = TooltipHelper.isShiftDown();
                     ClickType clickType = shiftDown ? ClickType.QUICK_MOVE : ClickType.PICKUP;
-                    SlotUtil.slotClickPhantom(slotReference, mouseButton, clickType, (ItemStack) ingredient);
+                    PhantomSlotUtil.slotClickPhantom(slotReference, mouseButton, clickType, (ItemStack) ingredient);
                     writeClientAction(1, buffer -> {
                         buffer.writeItemStack((ItemStack) ingredient);
                         buffer.writeVarInt(mouseButton);
@@ -117,7 +116,7 @@ public class PhantomSlotWidget extends SlotWidget implements IGhostIngredientTar
             int mouseButton = buffer.readVarInt();
             boolean shiftKeyDown = buffer.readBoolean();
             ClickType clickType = shiftKeyDown ? ClickType.QUICK_MOVE : ClickType.PICKUP;
-            SlotUtil.slotClickPhantom(slotReference, mouseButton, clickType, stackHeld);
+            PhantomSlotUtil.slotClickPhantom(slotReference, mouseButton, clickType, stackHeld);
         } else if (id == 2) {
             slotReference.putStack(ItemStack.EMPTY);
         }
